refactor(lowy): extract CommentIframeSurface mirroring CommentTextSurface
Lowy #1: text browse and branch diff wrap their Pierre viewer in
`CommentTextSurface` so the capture mechanism is fully encapsulated.
HTML iframe had three `createEffect` calls in `BrowsePreviewView`
doing the equivalent inline — symmetry violation. Extract a
`CommentIframeSurface` sibling so the three capture surfaces have
parallel structure. Future iframe consumers slot in via the wrapper
rather than re-implementing the bridge wiring.

Also collapses Hickey #1: the dead `pushPathTo` effect (`props.filePath`
is immutable per BrowsePreviewView lifetime; the bridge's `onLoad` handler
already pushes path on every iframe load) disappears with the rewrite.
